+++ src/Transformers/StopWordFilter.php
@@ -25,17 +25,12 @@ class StopWordFilter extends RegexFilter
     public function __construct(array $stopWords = [])
     {
         foreach ($stopWords as &$word) {
-            if (!is_string($word) || empty($word)) {
+            if (!is_string($word) or empty($word)) {
                 throw new InvalidArgumentException('Stop word must be a'
                     . 'non-empty string, ' . gettype($word) . ' found.');
             }
 
             $word = preg_quote($word, '/');
-        }
-
-        if (empty($stopWords)) {
-            throw new InvalidArgumentException('Must specify at least'
-                . ' 1 stop word');
         }
 
         $pattern = sprintf('/\b(%s)\b/u', implode('|', $stopWords));