+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 - 0.2.1
+    - Optimized Stop Word Filter
+    - Allow list of empty regex patterns in Regex Filter
     - Handle missing class definitions in Native and Igbinary
 
 - 0.2.0